SVN revision r1460
Documentation updates. Mostly this was getting the LDAP Scope handler to work in CILogon. A few other minor tweaks, such as updating the maven dependency plugin to 2.10 from 2.8 to allow for explicitly excluding certain dependencies. This was because one dependency had a dependency on a Bouncy Castle jar which promptly gave Tomcat some sort of cyclical error trying to load it. This has been rectified. Fix for the OA2 client servlet to properly show the redirect url if so enabled.