Number of fixes to prevent stacktrace to user
A number of exceptions can be replaced with e.g. a bad request to the user by
throwing a OA2GeneralError instead of a GeneralException. This was already
(partially) done in OA2CertServlet.java, not yet in OA2ATServlet.java. There are
probably more...
Also, a UnknownClientException should be returned using a OA2GeneralError()
(which does not use the callback URI) not via a ServletException() which would
result in a stacktrace.